refactor: import overlay symbols from barrel exports
Replace deep imports from OverlayDrawTarget, types, and layoutPlan
with the ../overlay barrel in BTAPI, renderers, and tests.
