+++ src/core/BTAPI.ts
@@ -13,8 +13,8 @@ import { createSystemFont } from '../assets/SystemFont';
 import { GamepadInput } from '../input/GamepadInput';
 import { KeyboardInput } from '../input/KeyboardInput';
 import { PointerInput } from '../input/PointerInput';
+import type { OverlayDrawTarget } from '../overlay';
 import { createOverlayLayout, Overlay, resolveOverlayTopLeftLabel } from '../overlay';
-import type { OverlayDrawTarget } from '../overlay/OverlayDrawTarget';
 import type { Effect } from '../render/effects/Effect';
 import type { IRenderer } from '../render/IRenderer';
 import { SoftwareRenderer } from '../render/SoftwareRenderer';
